feature: trace support
