--- pkg/trace/otlp/otlp.go
@@ -1,7 +1,0 @@
-package otlp
-
-//
-//func NewOTLPExporter(ctx context.Context, cfg *model.TracerConfig) (sdktrace.SpanExporter, error) {
-//	client := otlptracehttp.NewClient()
-//	return otlptrace.New(ctx, client)
-//}
