+++ pkg/tracing/driver.go
@@ -1,25 +1,51 @@
-package trace
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package tracing
 
 import (
 	"context"
-	"github.com/apache/dubbo-go-pixiu/pkg/model"
-	"github.com/apache/dubbo-go-pixiu/pkg/trace/jaeger"
-	"github.com/apache/dubbo-go-pixiu/pkg/trace/otlp"
+)
+
+import (
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/sdk/resource"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	semconv "go.opentelemetry.io/otel/semconv/v1.7.0"
 )
 
-// 不同listener监听的协议请求不同
+import (
+	"github.com/apache/dubbo-go-pixiu/pkg/model"
+	"github.com/apache/dubbo-go-pixiu/pkg/tracing/jaeger"
+	"github.com/apache/dubbo-go-pixiu/pkg/tracing/otlp"
+)
+
+// Different Listeners listen to different protocol requests, and create the corresponding tracer
 type ProtocolName string
 
 const HTTP ProtocolName = "HTTP"
 
+// Unique Name by making Id self-incrementing。
 type Holder struct {
 	Tracers map[string]Trace
 	Id      uint64
 }
+
+// Tracers corresponding to the listening protocol are maintained by the holder
 type TraceDriver struct {
 	Holders map[ProtocolName]*Holder
 	Tp      *sdktrace.TracerProvider
@@ -32,6 +58,7 @@ func NewTraceDriver() *TraceDriver {
 	}
 }
 
+// Loading BootStrap configuration about trace
 func (driver *TraceDriver) Init(bs *model.Bootstrap) *TraceDriver {
 	config := bs.Trace
 	ctx := context.Background()
@@ -70,7 +97,7 @@ func newTraceProvider(exp sdktrace.SpanExporter, cfg *model.TracerConfig) *sdktr
 	)
 }
 
-// 默认是fraction
+// Default sampler's Type: fraction
 func newSampler(sample model.Sampler) sdktrace.Sampler {
 	if sample.Type == "never" {
 		return sdktrace.NeverSample()
